[3957] Make build system more customizable (!8)
add existing check before unlinking

[3957] Make build system more customizable

"--compiler" option let user to specify desired compiler. NB: there are no any checks of user input; all the responsibility is on user!
"--cbackend" option is a switch from llc to llvm-cbe, that may be used in order to compile on/for systems that are not supported by llc

Reviewed-on: https://dsfgit.devos.club/antiq/typescript-llvm/pulls/8
Reviewed-by: alex.voloshanenko <alexander.voloshanenko@myoffice.team>