[AN-451] Workaround specific case of interface types in function signatures (!72)
Interfaces with first class functions as signature return type are not allowed for now because of environment-related implementation specifics. This is a workaround that allows such cases only for function declarations and not pretend to be a proper solution. Environment lifecycle should be revised to make it works as it should.

Reviewed-on: https://gitea.devos.club/antiq/typescript-llvm/pulls/72
Reviewed-by: alex.voloshanenko <alexander.voloshanenko@myoffice.team>
Co-authored-by: eugene.makarov <eugene.makarov@myoffice.team>
Co-committed-by: eugene.makarov <eugene.makarov@myoffice.team>